Add files via upload
A non- linear regression model using Cubic Spline is employed to predict the number of rings on abalones, Whole weight is used as a predictor and Rings as the target variable. A cubic Spline with two knots at the 33rd and and 67th percentile is used and the Root Mean Square Logarithmic Error (RMSLE)= 0.1948 is obtained and results submitted to Kaggle as a late submission.